Separate RTCStats on different file
